requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
openai
pypdf
PyMuPDF
matplotlib
pandas
playwright
httpx
pydantic
pypdf
PyMuPDF
matplotlib
pandas
playwright
